+++ src/database/src/Commands/Ast/ModelUpdateVisitor.php
@@ -32,11 +32,8 @@ use Hyperf\Utils\Str;
 use PhpParser\Comment\Doc;
 use PhpParser\Node;
 use PhpParser\NodeVisitorAbstract;
-use Roave\BetterReflection\BetterReflection;
 use Roave\BetterReflection\Reflection\ReflectionClass;
 use Roave\BetterReflection\Reflection\ReflectionMethod;
-use Roave\BetterReflection\Reflector\ClassReflector;
-use Roave\BetterReflection\TypesFinder\FindReturnType;
 use RuntimeException;
 
 class ModelUpdateVisitor extends NodeVisitorAbstract
@@ -79,16 +76,6 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
      * @var array
      */
     protected $properties = [];
-
-    /**
-     * @var ClassReflector
-     */
-    protected static $reflector;
-
-    /**
-     * @var FindReturnType
-     */
-    protected static $return;
 
     public function __construct($class, $columns, ModelOption $option)
     {
@@ -211,7 +198,7 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
     protected function initPropertiesFromMethods()
     {
         /** @var ReflectionClass $reflection */
-        $reflection = self::getReflector()->reflect(get_class($this->class));
+        $reflection = BetterReflectionManager::getReflector()->reflect(get_class($this->class));
         $methods = $reflection->getImmediateMethods();
         $namespace = $reflection->getDeclaringNamespaceAst();
         $casts = $this->class->getCasts();
@@ -223,8 +210,15 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
                 // Magic get<name>Attribute
                 $name = Str::snake(substr($method->getName(), 3, -9));
                 if (! empty($name)) {
-                    $type = self::getReturnFinder()->__invoke($method, $namespace);
-                    $this->setProperty($name, $type, true, null);
+                    $type = BetterReflectionManager::getReturnFinder()->__invoke($method, $namespace);
+                    if (empty($type) && $returnType = $method->getReturnType()) {
+                        $returnTypeName = $returnType->getName();
+                        if (class_exists($returnTypeName)) {
+                            $returnTypeName = '\\' . $returnTypeName;
+                        }
+                        $type = [$returnTypeName];
+                    }
+                    $this->setProperty($name, $type, true, null, '', false, 1);
                 }
                 continue;
             }
@@ -233,7 +227,7 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
                 // Magic set<name>Attribute
                 $name = Str::snake(substr($method->getName(), 3, -9));
                 if (! empty($name)) {
-                    $this->setProperty($name, null, null, true);
+                    $this->setProperty($name, null, null, true, '', false, 1);
                 }
                 continue;
             }
@@ -297,7 +291,7 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
             }
 
             if (is_subclass_of($caster, CastsAttributes::class)) {
-                $ref = self::getReflector()->reflect($caster);
+                $ref = BetterReflectionManager::getReflector()->reflect($caster);
                 $method = $ref->getMethod('get');
                 if ($ast = $method->getReturnStatementsAst()[0]) {
                     if ($ast instanceof Node\Stmt\Return_
@@ -311,7 +305,7 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
         }
     }
 
-    protected function setProperty(string $name, array $type = null, bool $read = null, bool $write = null, string $comment = '', bool $nullable = false)
+    protected function setProperty(string $name, array $type = null, bool $read = null, bool $write = null, string $comment = '', bool $nullable = false, int $priority = 0)
     {
         if (! isset($this->properties[$name])) {
             $this->properties[$name] = [];
@@ -319,6 +313,10 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
             $this->properties[$name]['read'] = false;
             $this->properties[$name]['write'] = false;
             $this->properties[$name]['comment'] = (string) $comment;
+            $this->properties[$name]['priority'] = 0;
+        }
+        if ($this->properties[$name]['priority'] > $priority) {
+            return;
         }
         if ($type !== null) {
             if ($nullable) {
@@ -332,6 +330,7 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
         if ($write !== null) {
             $this->properties[$name]['write'] = $write;
         }
+        $this->properties[$name]['priority'] = $priority;
     }
 
     protected function setMethod(string $name, array $type = [], array $arguments = [])
@@ -408,23 +407,5 @@ class ModelUpdateVisitor extends NodeVisitorAbstract
         /** @var Model $model */
         $model = new $className();
         return '\\' . get_class($model->newCollection());
-    }
-
-    protected static function getReturnFinder(): FindReturnType
-    {
-        if (static::$return instanceof FindReturnType) {
-            return static::$return;
-        }
-
-        return static::$return = new FindReturnType();
-    }
-
-    protected static function getReflector(): ClassReflector
-    {
-        if (self::$reflector instanceof ClassReflector) {
-            return self::$reflector;
-        }
-
-        return self::$reflector = (new BetterReflection())->classReflector();
     }
 }