+++ src/database/src/Model/Concerns/CamelCase.php
@@ -34,6 +34,15 @@ trait CamelCase
         return $array;
     }
 
+    public function getFillable()
+    {
+        $fillable = [];
+        foreach (parent::getFillable() as $key) {
+            $fillable[] = $this->keyTransform($key);
+        }
+        return $fillable;
+    }
+
     public function toArray(): array
     {
         $array = [];
@@ -51,19 +60,5 @@ trait CamelCase
     protected function keyTransform($key)
     {
         return Str::camel($key);
-    }
-
-    protected function addMutatedAttributesToArray(array $attributes, array $mutatedAttributes)
-    {
-        foreach ($mutatedAttributes as $key) {
-            if (! array_key_exists($this->keyTransform($key), $attributes)) {
-                continue;
-            }
-            $attributes[$this->keyTransform($key)] = $this->mutateAttributeForArray(
-                $this->keyTransform($key),
-                $attributes[$this->keyTransform($key)]
-            );
-        }
-        return $attributes;
     }
 }