+++ src/redis/src/Redis.php
@@ -57,6 +57,7 @@ class Redis
                     // Should storage the connection to coroutine context, then use defer() to release the connection.
                     Context::set($this->getContextKey(), $connection);
                     defer(function () use ($connection) {
+                        Context::set($this->getContextKey(), null);
                         $connection->release();
                     });
                 } else {