+++ src/websocket-server/tests/ContextTest.php
@@ -9,14 +9,17 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
+namespace HyperfTest\WebSocketServer;
+
 use Hyperf\Utils\Context as CoContext;
 use Hyperf\WebSocketServer\Context;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @internal
  * @coversNothing
  */
-class ContextTest extends \PHPUnit\Framework\TestCase
+class ContextTest extends TestCase
 {
     public function testHas()
     {